+++ subsys/bluetooth/fast_pair/fhn/motion_detector.c
@@ -11,7 +11,7 @@ LOG_MODULE_REGISTER(fp_fhn_motion_detector, CONFIG_BT_FAST_PAIR_LOG_LEVEL);
 
 #include <bluetooth/fast_pair/fast_pair.h>
 #include <bluetooth/fast_pair/fhn/fhn.h>
-#include <dult.h>
+#include <dult/dult.h>
 
 #include "fp_fhn_dult_integration.h"
 #include "fp_activation.h"
